Fix tooth width.
I forgot that the division between tooth and space has to be exactly equal at the Teilkreis (not sure what the precise english term is), not at the base circle. To achieve this, the tooth has to be wider at the base, which means the angle needs to be increased by whatever angle is necessary to get from the intersection base circle/involute to the intersection teilkreis/involute.